openapi: add routeWithOpenApi
This change adds a routeWithOpenApi method to the base controller
class that we use. That method takes modifies the open api
configuration passed in and adds standard error codes. It uses that
method in the addon controller.

The goal here is to simplify how we add error codes to the openapi
doc, potentially removing the need for us to do that entirely.

This is just a proof of concept to show that it works. The API for it
could use some work (for instance, having to pass in the service is a
little inconvenient) and possibly some tests, but this shows that
it **is** possible. Might be a nice QOL improvement.