feat(#4205): allow all addons to listen for potentially stale events
This change lets all addons listen for events when features get marked
or unmarked as potentially stale.

## Discussion

### All addons?

Should this be available to all addons? I can't see a reason why it
shouldn't be available to all addons, but I might be missing
something.

### Should it be behind a flag?

The feature is still behind a flag, but the event type is not. Should
we gate the event being available until we actually emit the event?
That would require some more code, but could yield less potential confusion.

Open to hearing your thoughts.