fix: EventStore#getMaxRevisionId can return null
In a new fresh Unleash instance with cache enabled this can cause
feature toggles to never get updated.

We saw in our client that the ETag was ETag: "60e35fba:null"
Which looked incorrect for us.

I also did manual testing and if the andWhere had a value of
largerThan higher than whatever the id was then we would get
back { max: null }.

This should fix that issue.