fix: CR strategy name changes code
This change addresses two things that were done in
https://github.com/Unleash/unleash/pull/4004 and that I believe to be bugs.

1. It shows the previous strategy name also if there was no previous
title. So if there was no previous title, it'll show the strategy name
with a strikethrough and then the new title.
2. It changes a `span` component to a `del` component. I believe the
span was erroneously changed from a `s` component (strikethrough
component) in the linked PR (based on a comment on the PR). This
caused the strikethrough to not be there anymore. However, the `del`
component is semantically more correct and reintroduces the
strikethrough, so it is a better change.

Finally, it removes a redundant pair of curly braces.