chore: pin node version 18.17.1
We are pinning node to version 18.17.1 as we have seen some performance degregation on node 18.18.0 on arm64. We will investigate this further at a later point. This is to mitigate the issue.

Our next step in pinpointing the issue will be to compare between running on musl vs libc on arm64.