# fix(docs): make docs build not break on md links with uppercases
Markdown generators (including GitHub) tend to lowercase the titles
when they create anchor links. Most browsers might be able to handle
mixed case, but some of our tools don't.

This is currently breaking the docs build.

This fixes the issue by lowercasing any links we find to internal headers.

## Discussion point

Now, there is one potential issue with this: if someone creates an explicit link in the SDKs that uses uppercase letters, then this might break the docs build in the future.

However, I think this is unlikely to happen any time soon, and I would think that it's more likely that people will incorrectly case the header link as happened in the Ruby repo.