fix(docs): lowercase internal markdown links (#4963)
Markdown generators (including GitHub) tend to lowercase the titles
when they create anchor links. It appears that the intra-doc links might
not work correctly if they're incorrectly cased.

This fixes the issue by lowercasing any links we find to internal
headers in external documents (such as the SDKs and Edge/Proxy docs).

## Discussion point

Now, there is one potential issue with this: if someone creates an
explicit link in the SDKs that uses uppercase letters, then this might
break the docs build in the future.

However, I think this is unlikely to happen any time soon, and I would
think that it's more likely that people will incorrectly case the header
link.